+++ src/ast/passes/field_analyser.cpp
@@ -247,13 +247,8 @@ void FieldAnalyser::resolve_args(Probe &probe)
               func, probe_type == ProbeType::fexit, true, false, err);
 
         } else if (probe_type == ProbeType::rawtracepoint) {
-          for (const auto &prefix : RT_BTF_PREFIXES) {
-            maybe_ap_args = bpftrace_.btf_->resolve_args(
-                prefix + ap->func, false, false, true, err);
-            if (maybe_ap_args.has_value()) {
-              break;
-            }
-          }
+          maybe_ap_args = bpftrace_.btf_->resolve_raw_tracepoint_args(func,
+                                                                      err);
         }
 
         if (!maybe_ap_args.has_value()) {
@@ -278,13 +273,8 @@ void FieldAnalyser::resolve_args(Probe &probe)
         maybe_probe_args = bpftrace_.btf_->resolve_args(
             ap->func, probe_type == ProbeType::fexit, true, false, err);
       } else if (probe_type == ProbeType::rawtracepoint) {
-        for (const auto &prefix : RT_BTF_PREFIXES) {
-          maybe_probe_args = bpftrace_.btf_->resolve_args(
-              prefix + ap->func, false, false, true, err);
-          if (maybe_probe_args.has_value()) {
-            break;
-          }
-        }
+        maybe_probe_args = bpftrace_.btf_->resolve_raw_tracepoint_args(ap->func,
+                                                                       err);
       }
 
       if (!maybe_probe_args.has_value()) {