+++ src/probe_matcher.h
@@ -99,9 +99,6 @@ private:
   virtual std::unique_ptr<std::istream> get_symbols_from_list(
       const std::vector<ProbeListItem> &probes_list) const;
 
-  virtual std::unique_ptr<std::istream> adjust_rawtracepoint(
-      std::istream &symbol_list) const;
-
   std::unique_ptr<std::istream> get_iter_symbols() const;
 
   std::unique_ptr<std::istream> kernel_probe_list();