+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to
   - [#3905](https://github.com/bpftrace/bpftrace/pull/3905)
 - Rawtracepoints can now use `args` builtin and list params
   - [#3918](https://github.com/bpftrace/bpftrace/pull/3918)
+- Add ability to specify rawtracepoint modules
+  - [#3944](https://github.com/bpftrace/bpftrace/pull/3944)
 #### Changed
 - `-p` CLI flag now applies to all probes (except BEGIN/END)
   - [#3800](https://github.com/bpftrace/bpftrace/pull/3800)